Refactor into ShapeGeometry class
