+++ src/ShapeCrawler/ShapeCollection/GroupShapes/GroupedShape.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Linq;
+﻿using System.Linq;
 using DocumentFormat.OpenXml;
 using ShapeCrawler.Shared;
 using P = DocumentFormat.OpenXml.Presentation;
@@ -133,7 +132,15 @@ internal sealed class GroupedShape : IShape
     
     public PlaceholderType PlaceholderType => this.decoratedShape.PlaceholderType;
    
-    public Geometry GeometryType => this.decoratedShape.GeometryType;
+    public Geometry GeometryType {
+        get => this.decoratedShape.GeometryType;
+        set => this.decoratedShape.GeometryType = value;
+    }
+
+    public decimal? CornerSize {
+        get => this.decoratedShape.CornerSize;
+        set => this.decoratedShape.CornerSize = value;
+    }
 
     public string? CustomData
     {
@@ -167,12 +174,6 @@ internal sealed class GroupedShape : IShape
     {
         get => this.TextBox.Text;
         set => this.TextBox.Text = value;
-    }
-
-    public decimal? CornerSize
-    {
-        get => null;
-        set => throw new NotImplementedException();
     }
 
     public void Remove() => this.decoratedShape.Remove();
