+++ src/ShapeCrawler/ShapeCollection/IShape.cs
@@ -7,7 +7,7 @@ namespace ShapeCrawler;
 /// <summary>
 ///     Represents a shape.
 /// </summary>
-public interface IShape : IPosition
+public interface IShape : IPosition, IShapeGeometry
 {
     /// <summary>
     ///     Gets or sets the width of the shape in pixels.
@@ -48,21 +48,6 @@ public interface IShape : IPosition
     ///     Gets the placeholder type of the shape.
     /// </summary>
     PlaceholderType PlaceholderType { get; }
-
-    /// <summary>
-    ///     Gets the geometry form of the shape.
-    /// </summary>
-    Geometry GeometryType { get; }
-
-    /// <summary>
-    ///     Gets or sets the size of the corners in this shape. Range: 0.0-1.0.
-    /// </summary>
-    /// <remarks>
-    ///     Applicable only to rounded rectangle geometry.
-    ///     Value 0.0 makes shape appear as a regular rectangle with no corners.
-    ///     Value 1.0 is the maximum size of a corner: 1/2 length of shortest side.
-    /// </remarks>
-    decimal? CornerSize { get; set; }
 
     /// <summary>
     ///     Gets or sets custom data string for the shape.
