+++ src/ShapeCrawler/ShapeCollection/Shape.cs
@@ -6,7 +6,6 @@ using DocumentFormat.OpenXml.Packaging;
 using ShapeCrawler.Exceptions;
 using ShapeCrawler.Extensions;
 using ShapeCrawler.Texts;
-using A = DocumentFormat.OpenXml.Drawing;
 using P = DocumentFormat.OpenXml.Presentation;
 using Position = ShapeCrawler.Positions.Position;
 
@@ -142,50 +141,16 @@ internal abstract class Shape : IShape
         }
     } 
         
-    public virtual Geometry GeometryType => Geometry.Rectangle;
-
-    public decimal? CornerSize
+    public virtual Geometry GeometryType
     {
-        get
-        {
-            var spPr = this.PShapeTreeElement.Descendants<P.ShapeProperties>().First();
-            var aPresetGeometry = spPr.GetFirstChild<A.PresetGeometry>();
-            var shapeType = aPresetGeometry?.Preset?.Value;
-
-            if (shapeType == A.ShapeTypeValues.RoundRectangle)
-            {
-                return ExtractCornerSizeFromRoundRectangle(aPresetGeometry!);
-            }
-
-            if (shapeType == A.ShapeTypeValues.Round2SameRectangle)
-            {
-                return ExtractCornerSizeFromRound2SameRectangle(aPresetGeometry!);
-            }
-
-            return null;
-        }
-        
-        set
-        {
-            if (value is null)
-            {
-                throw new SCException("Not allowed to set null size. Try 0 to straighten the corner.");
-            }
-
-            var spPr = this.PShapeTreeElement.Descendants<P.ShapeProperties>().First();
-            var aPresetGeometry = spPr.GetFirstChild<A.PresetGeometry>();
-            var shapeType = aPresetGeometry?.Preset?.Value;
-
-            if (shapeType == A.ShapeTypeValues.RoundRectangle)
-            {
-                InjectCornerSizeIntoRoundRectangle(aPresetGeometry!, value.Value);
-            }
-
-            if (shapeType == A.ShapeTypeValues.Round2SameRectangle)
-            {
-                InjectCornerSizeIntoRound2SameRectangle(aPresetGeometry!, value.Value);
-            }
-        }
+        get => Geometry.Rectangle;
+        set => throw new SCException("Changing geometry of this shape is not supported");
+    }
+     
+    public virtual decimal? CornerSize
+    {
+        get => null;
+        set => throw new SCException("Changing corner size of this shape is not supported");
     }
 
     public string? CustomData
@@ -275,127 +240,4 @@ internal abstract class Shape : IShape
             $"The shape is not a media shape. Use {nameof(IShape.ShapeType)} property to check if the shape is a media (audio, video, etc.");
     
     public virtual void Remove() => this.PShapeTreeElement.Remove();
-
-    private static decimal? ExtractCornerSizeFromRoundRectangle(A.PresetGeometry aPresetGeometry)
-    {
-        if (aPresetGeometry.Preset?.Value != A.ShapeTypeValues.RoundRectangle)
-        {
-            return null;
-        }
-
-        var avList = aPresetGeometry.AdjustValueList ?? throw new SCException("Malformed rounded rectangle. Missing AdjustValueList. Please file a GitHub issue.");
-        var sgs = avList.Descendants<A.ShapeGuide>().Where(x => x.Name == "adj");
-        if (!sgs.Any())
-        {
-            // Has no shape guide. That means we're using the DEFAULT value, which is 0.35
-            return 0.35m;
-        }
-
-        if (sgs.Count() > 1)
-        {
-            throw new SCException("Malformed rounded rectangle. Has multiple shape guides. Please file a GitHub issue.");
-        }
-
-        return ExtractCornerSizeFromShapeGuide(sgs.Single());
-    }
-
-    private static void InjectCornerSizeIntoRoundRectangle(A.PresetGeometry aPresetGeometry, decimal value)
-    {
-        if (aPresetGeometry.Preset?.Value != A.ShapeTypeValues.RoundRectangle)
-        {
-            return;
-        }
-
-        var avList = aPresetGeometry.AdjustValueList ?? throw new SCException("Malformed rounded rectangle. Missing AdjustValueList. Please file a GitHub issue.");
-        var sgs = avList.Descendants<A.ShapeGuide>().Where(x => x.Name == "adj");
-        if (sgs.Count() > 1)
-        {
-            throw new SCException("Malformed rounded rectangle. Has multiple shape guides. Please file a GitHub issue.");
-        }
-
-        // Will add a shape guide if there isn't already one
-        var sg = sgs.SingleOrDefault()
-            ?? avList.AppendChild(new A.ShapeGuide() { Name = "adj" }) 
-            ?? throw new SCException("Failed attempting to add a shape guide to AdjustValueList");
-
-        sg.Formula = new StringValue($"val {(int)(value * 50000m)}");        
-    }
-
-    private static decimal? ExtractCornerSizeFromRound2SameRectangle(A.PresetGeometry aPresetGeometry)
-    {
-        if (aPresetGeometry.Preset?.Value != A.ShapeTypeValues.Round2SameRectangle)
-        {
-            return null;
-        }
-
-        var avList = aPresetGeometry.AdjustValueList ?? throw new SCException("Malformed rounded rectangle. Missing AdjustValueList. Please file a GitHub issue.");
-        var sgs = avList.Descendants<A.ShapeGuide>();
-        var count = sgs.Count();
-        if (count == 0)
-        {
-            // Has no shape guide. That means we're using the DEFAULT value, which is 0.35
-            return 0.35m;
-        }
-
-        if (count != 2)
-        {
-            throw new SCException($"Malformed rounded rectangle. Expected 2 shape guides, found {count}. Please file a GitHub issue.");
-        }
-
-        var sg = sgs.SingleOrDefault(x => x.Name == "adj1") ?? throw new SCException($"Malformed rounded rectangle. No shape guide named `adj1`. Please file a GitHub issue.");
-
-        return ExtractCornerSizeFromShapeGuide(sg);
-    }
-
-    private static void InjectCornerSizeIntoRound2SameRectangle(A.PresetGeometry aPresetGeometry, decimal value)
-    {
-        if (aPresetGeometry.Preset?.Value != A.ShapeTypeValues.Round2SameRectangle)
-        {
-            return;
-        }
-
-        var avList = aPresetGeometry.AdjustValueList ?? throw new SCException("Malformed rounded rectangle. Missing AdjustValueList. Please file a GitHub issue.");
-        var sgs = avList.Descendants<A.ShapeGuide>().Where(x => x.Name == "adj1");
-        if (sgs.Count() > 1)
-        {
-            throw new SCException("Malformed rounded rectangle. Has multiple shape guides. Please file a GitHub issue.");
-        }
-
-        var sg = sgs.SingleOrDefault();
-        if (sg is null)
-        {
-            // Has no adj1 shape guide. We need to add an adj1/adj2 pair
-            sg = avList.AppendChild(new A.ShapeGuide() { Name = "adj1" }) ?? throw new SCException("Failed attempting to add a shape guide to AdjustValueList");
-            if (avList.AppendChild(new A.ShapeGuide() { Name = "adj2", Formula = "val 0" }) is null)
-            {
-                throw new SCException("Failed attempting to add a shape guide to AdjustValueList");
-            }
-        }
-    
-        sg.Formula = new StringValue($"val {(int)(value * 50000m)}");        
-    }
-
-    private static decimal ExtractCornerSizeFromShapeGuide(A.ShapeGuide sg)
-    {
-        var formula = sg.Formula?.Value ?? throw new SCException("Malformed rounded rectangle. Shape guide has no formula. Please file a GitHub issue.");
-
-        var pattern = "^val (?<value>[0-9]+)$";
-
-#if NETSTANDARD2_0
-        var regex = new Regex(pattern, RegexOptions.None, TimeSpan.FromSeconds(100));
-#else
-        var regex = new Regex(pattern, RegexOptions.NonBacktracking);
-#endif
-
-        var match = regex.Match(formula);
-        if (!match.Success)
-        {
-            throw new SCException("Malformed rounded rectangle. Formula has no value. Please file a GitHub issue.");
-        }
-
-        var value = int.Parse(match.Groups["value"].Value);
-
-        return value / 50000m;
-    }
-
 }
